requirements.txt
streamlit
rembg
pillow
onnxruntime